pyproject.toml
[project]
name = "personalplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.17.0,<2.0.0",
    "absolufy-imports==0.3.1",
    "alembic==1.17.1",
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "authlib==1.6.5",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cloudpickle==3.1.1",
    "comm==0.2.3",
    "cryptography==46.0.3",
    "dateparser==1.2.2",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "docstring-parser==0.17.0",
    "executing==2.2.1",
    "fastapi==0.120.4",
    "google-api-core==2.28.1",
    "google-api-python-client==2.186.0",
    "google-auth==2.42.1",
    "google-auth-httplib2==0.2.1",
    "google-auth-oauthlib==1.2.2",
    "google-cloud-aiplatform==1.124.0",
    "google-cloud-appengine-logging==1.7.0",
    "google-cloud-audit-log==0.4.0",
    "google-cloud-bigquery==3.38.0",
    "google-cloud-bigtable==2.34.0",
    "google-cloud-core==2.5.0",
    "google-cloud-discoveryengine==0.13.12",
    "google-cloud-logging==3.12.1",
    "google-cloud-monitoring==2.28.0",
    "google-cloud-resource-manager==1.15.0",
    "google-cloud-secret-manager==2.25.0",
    "google-cloud-spanner==3.59.0",
    "google-cloud-speech==2.34.0",
    "google-cloud-storage>=3.0.0,<4.0.0",
    "google-cloud-trace==1.17.0",
    "google-crc32c==1.7.1",
    "google-genai==1.47.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.71.0",
    "graphviz==0.21",
    "grpc-google-iam-v1==0.14.3",
    "grpc-interceptor==0.15.4",
    "grpcio==1.76.0",
    "grpcio-status==1.76.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "ipykernel==7.1.0",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "matplotlib-inline==0.2.1",
    "mcp==1.20.0",
    "nest-asyncio==1.6.0",
    "numpy==2.3.4",
    "oauthlib==3.3.1",
    "opentelemetry-api==1.37.0",
    "opentelemetry-exporter-gcp-logging==1.10.0a0",
    "opentelemetry-exporter-gcp-monitoring==1.10.0a0",
    "opentelemetry-exporter-gcp-trace==1.10.0",
    "opentelemetry-exporter-otlp-proto-common==1.37.0",
    "opentelemetry-exporter-otlp-proto-http==1.37.0",
    "opentelemetry-proto==1.37.0",
    "opentelemetry-resourcedetector-gcp==1.10.0a0",
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-semantic-conventions==0.58b0",
    "packaging==25.0",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "platformdirs==4.5.0",
    "prompt-toolkit==3.0.52",
    "proto-plus==1.26.1",
    "protobuf==6.33.0",
    "psutil==7.1.2",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "regex==2025.10.23",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "shapely==2.1.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "sqlalchemy-spanner==1.17.1",
    "sqlparse==0.5.3",
    "sse-starlette==3.0.3",
    "stack-data==0.6.3",
    "starlette==0.49.1",
    "tenacity>=9.0.0,<10.0.0",
    "tornado==6.5.2",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchdog==6.0.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "zipp==3.23.0",
    "streamlit>=1.51.0",
    "ngrok>=1.4.0",
]
[tool.uv]
managed = true

[tool.uv.sources]
google-adk = { git = "https://github.com/google/adk-python.git", rev = "main" }

[project.scripts]
dev = "dev:main"

[tool.setuptools]
packages = ["orchestrator", "calendar_service", "google_docs_service"]

